Renamed Autonomous Class to BeaconScorer, and improved Autonomous Programs
Renamed and refactored the Autonomous Class to the BeaconScorer to
better reflect it's purpose, clarify usage, and bring it inline with the
styles of the other robot class files. Also re-added back in the color
sensor to the autonomous program files (via the BeaconScorer class) with
improved functionality. Also tweaked some autonomous program constants).