Added BeaconTest class, improved ColorTest class, and removed deprecated ColorSensorTest class
Added a BeaconTest class that can be used to debug issues regarding the
Beacon Scorer. Removed redundant code from the ColorSensorTest class and
ColorTest class that was already done by the MRRGBExample class. Did so
by removing the ColorSensorTest class, and changing the ColorTest class
to show the what BeaconScorer recognizes from the beacon.